tmkms-p2p: extract `framing` module
This renames the `protocol` module to `framing`, and ensures it only
contains functions for encoding/decoding message frames.
